Moved io folder to a file_parsers folder under parsers
