+++ aiida_vasp/parsers/file_parsers/chgcar.py
@@ -1,6 +1,6 @@
 """Tools for parsing CHGCAR files."""
 
-from aiida_vasp.io.parser import BaseFileParser
+from aiida_vasp.parsers.file_parsers.parser import BaseFileParser
 from aiida_vasp.parsers.node_composer import NodeComposer
