+++ aiida_vasp/parsers/file_parsers/doscar.py
@@ -2,7 +2,7 @@
 import numpy as np
 
 from aiida_vasp.parsers.node_composer import NodeComposer
-from aiida_vasp.io.parser import BaseFileParser
+from aiida_vasp.parsers.file_parsers.parser import BaseFileParser
 
 # Map from number of columns in DOSCAR to dtype.
 DTYPES = {
