+++ aiida_vasp/parsers/file_parsers/incar.py
@@ -9,7 +9,7 @@ import pyparsing as pp
 import six
 from pymatgen.io.vasp import Incar as IncarPymatgen
 from parsevasp.incar import Incar as IncarParsevasp
-from aiida_vasp.io.parser import BaseFileParser
+from aiida_vasp.parsers.file_parsers.parser import BaseFileParser
 
 from aiida_vasp.utils.aiida_utils import get_data_node, get_data_class
