+++ aiida_vasp/parsers/file_parsers/kpoints.py
@@ -2,7 +2,7 @@
 """Utils for VASP KPOINTS format"""
 
 from parsevasp.kpoints import Kpoints, Kpoint
-from aiida_vasp.io.parser import BaseFileParser
+from aiida_vasp.parsers.file_parsers.parser import BaseFileParser
 from aiida_vasp.parsers.node_composer import NodeComposer
 from aiida_vasp.utils.aiida_utils import get_data_class
