+++ aiida_vasp/parsers/file_parsers/potcar.py
